requirements.txt
aiohttp
aiofiles
discord.py
requests
pillow
TwitterAPI